+++ include/client/env.hpp
@@ -60,6 +60,7 @@ static constexpr auto METRICS_IP_PORT = ADD_PREFIX("METRICS_IP_PORT");
 
 static constexpr auto NUM_REPL = ADD_PREFIX("NUM_REPL");
 static constexpr auto PROXY_PID_FILE = ADD_PREFIX("PROXY_PID_FILE");
+static constexpr auto DENTRY_CACHE = ADD_PREFIX("DENTRY_CACHE");
 
 } // namespace gkfs::env