+++ src/client/preload_util.cpp
@@ -32,6 +32,7 @@
 #include <client/logging.hpp>
 #include <client/rpc/forward_metadata.hpp>
 #include <client/rpc/forward_metadata_proxy.hpp>
+#include <client/cache.hpp>
 
 #include <common/rpc/distributor.hpp>
 #include <common/rpc/rpc_util.hpp>
@@ -45,6 +46,7 @@
 #include <regex>
 #include <csignal>
 #include <random>
+#include <filesystem>
 
 extern "C" {
 #include <sys/sysmacros.h>
@@ -209,6 +211,27 @@ optional<gkfs::metadata::Metadata>
 get_metadata(const string& path, bool follow_links) {
     std::string attr;
     int err{};
+    // Use file metadata from dentry cache if available
+    if(CTX->use_dentry_cache()) {
+        // get parent and filename path to retrieve the cache entry
+        std::filesystem::path p(path);
+        auto parent = p.parent_path().string();
+        auto filename = p.filename().string();
+        auto cache_entry = CTX->dentry_cache()->get(parent, filename);
+        if(cache_entry) {
+            LOG(DEBUG, "{}(): Dentry cache hit for file '{}'", __func__, path);
+            // if cache_entry exists, generate a Metadata object from it.
+            mode_t mode = gkfs::config::syscall::stat::file_mode_default;
+            if(cache_entry->file_type == gkfs::filemap::FileType::directory) {
+                mode = gkfs::config::syscall::stat::dir_mode_default;
+            }
+            gkfs::metadata::Metadata md{};
+            md.mode(mode);
+            md.ctime(cache_entry->ctime);
+            md.size(cache_entry->size);
+            return md;
+        }
+    }
     if(gkfs::config::proxy::fwd_stat && CTX->use_proxy()) {
         err = gkfs::rpc::forward_stat_proxy(path, attr);
     } else {